fix(app): can not import module from different directory
