--- server/server.py
@@ -1,35 +1,0 @@
-import grpc
-from concurrent import futures
-
-import services.grpc_generated.chat_pb2_grpc as chat_pb2_grpc
-import services.grpc_generated.user_pb2_grpc as user_pb2_grpc
-
-from services_class.user_service import UserServiceServicer 
-from services_class.chat_service import ChatServiceServicer 
-
-"""
-Main entry point for the gRPC server.
-"""
-def Server():
-	# create a gRPC server
-	server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
-	
-	# add servicers to the server
-	chat_pb2_grpc.add_ChatServiceServicer_to_server(
-		ChatServiceServicer(), server)
-	user_pb2_grpc.add_UserServiceServicer_to_server(
-		UserServiceServicer(), server)
-	
-	# listen on port 50051
-	server.add_insecure_port('[::]:50051')
-
-	# start gRPC server
-	server.start()
-	print("Server started")
-
-	# since server.start() will not block,
-	server.wait_for_termination()
-
-
-if __name__ == '__main__':
-	Server()
